% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_workfile}
\alias{prep_workfile}
\title{Prep Workfiles}
\usage{
prep_workfile(x, widths)
}
\description{
Adds spaces to meet the width requirements of each element. Replaces NA with appropriate amount of spaces.
}
