% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_xb}
\alias{prep_xb}
\title{Prep XB}
\usage{
prep_xb(xb)
}
\description{
Since the XB files essentially contains 3 different files,
 this file will clean each of the files read in by `read_xb` separately.
}
